+++ generate/graphql/root_tpl.go
@@ -3,6 +3,7 @@ package graphql
 const ROOT_TEMPLATE = `
 # java.util.Date implementation
 scalar Date
+scalar Long
 
 type Query {
 {{- range . }}